seat-select-backend/pyproject.toml
[project]
name = "seat-select-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "supabase>=2.27.0",
    "tzdata>=2025.3",
    "uvicorn>=0.40.0",
]
